:construction: Added Sessions, Updated UI
Updated the README file with some more information on the project, as well as a current to-do list of features I would like to add.

I added in session variables that keep track of whether the user is currently logged in, and redirects the user back to the log-in page if they're not.

Added in a fixed header with a logo and links for future pages. Currently, only the Log Out function actually does anything.

Changed the accent colors from purple to red. This was done because ELA and English classes are associated with the color red, so I figured that it will help stick to the minimal branding.