:sparkles: Updates to Log In / Account Creation
Added a password confirmation field to prevent users from submitting words with typos.

Added two PHP checks on submission that will block registration. The first checks to see if the username you requested already exists in the database. The second checks to see if the two password fields have matching values. If either one fails, the user will be booted back to the prompt catalog screen.

Added in a check to see if the user is already logged in to prevent them from accessing login.php and create-account.php

Updated the visuals on both the login and account creation pages so that they're a lot cleaner.

Changed some of the phrases used on these pages.