:lipstick: Smaller Homepage Details
Changed the text color in the search button from "white" to "ui-white-text-color" to keep it consistent with the rest of the interface.

Added a hover effect to the header logo text so that it shifts to the chosen accent color.

Removed the corresponding task from the To-Do list, and added the new task of organizing the folder structure.